% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasteps_translators.R
\name{DS_translate}
\alias{DS_translate}
\title{DataStep Wrapper}
\usage{
DS_translate(
  in_file = NULL,
  out_file = "dmcas_scorecode.R",
  zip = TRUE,
  hostname = NULL,
  in_caslib,
  in_castable,
  out_caslib,
  out_castable
)
}
\arguments{
\item{in_file}{Path to zip/sas file with the score code.}

\item{out_file}{Path to the translated \code{.R} code file.}

\item{zip}{Boolean, if the score code is inside a zip file. Default to \code{TRUE}. It will look for \code{dmcas_scorecode.sas} file inside the zip.}

\item{hostname}{\code{NULL} by default, extracts the server name from \code{dmcas_scorecode.sas} file.}

\item{in_caslib}{caslib name of the input table}

\item{in_castable}{input table name}

\item{out_caslib}{caslib name of the output table}

\item{out_castable}{output table name}
}
\value{
List object with the DS code, Rscore code, out castable, out caslib and the written file path.
}
\description{
Wraps a simple DataSetp code (not DS2) to be run through \code{SWAT}
}
\examples{
DS_translate("filepath.zip")

}
