% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasteps_translators.R
\name{EPS_translate}
\alias{EPS_translate}
\title{EPS Translator}
\usage{
EPS_translate(
  in_file = NULL,
  out_file = "dmcas_epscorecode.R",
  zip = TRUE,
  in_caslib,
  in_castable,
  out_caslib,
  out_castable,
  hostname = "sasserver.demo.sas.com",
  copyVars = NULL
)
}
\arguments{
\item{in_file}{Path to zip/sas file with the score code.}

\item{out_file}{Path to the translated \code{.R} code file.}

\item{zip}{Boolean, if the score code is inside a zip file. Default to \code{TRUE}. It will look for \code{dmcas_epscorecode.sas} file inside the zip.}

\item{in_caslib}{caslib name of the input table}

\item{in_castable}{input table name}

\item{out_caslib}{caslib name of the output table}

\item{out_castable}{output table name}

\item{hostname}{sas viya hostname to be used, not available inside the DS2 code}

\item{copyVars}{default \code{NULL}, will not copy variables to the output. If \code{"ALL"} will copy all variables to the scored table output, if it is a vector, will copy named vars e.g: \verb{c("var1", "var2)}}
}
\value{
List object with the Rscore code, out castable, out caslib and the written file path.
}
\description{
It will read the score code that is written in DS2, extract the astore name and
create an astore call written using \code{SWAT}. The reason for that is because the DS2 is
not reliable/self contained
}
\examples{
EPS_translate("filepath.zip")

}
