% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_translators.R
\name{nlp_sentiment_translate}
\alias{nlp_sentiment_translate}
\title{NLP Sentiment Generator/Translator}
\usage{
nlp_sentiment_translate(
  in_file = NULL,
  out_file = "SentimentScoreCode.R",
  zip = TRUE,
  in_caslib,
  in_castable,
  out_caslib,
  out_castable_sentiment,
  out_castable_matches,
  out_castable_features,
  key_column,
  document_column,
  hostname = NULL,
  astore = FALSE,
  astore_caslib = "casuser",
  astore_name = "Sentiment_Astore",
  astore_path = "SentimentModel.astore",
  copyVars = NULL
)
}
\arguments{
\item{in_file}{Path to zip/sas file with the score code.}

\item{out_file}{Path to the translated \code{.R} code file.}

\item{zip}{Boolean, if the score code is inside a zip file. Default to \code{TRUE}. It will look for \code{dmcas_epscorecode.sas} file inside the zip.}

\item{in_caslib}{caslib name of the input table}

\item{in_castable}{input table name}

\item{out_caslib}{caslib name of the output table}

\item{out_castable_sentiment}{sentiment output table name}

\item{out_castable_matches}{matches output table name, if the argument is not defined the table name will be the same as \code{out_castable_sentiment} with "_matches" added}

\item{out_castable_features}{features output table name, if the argument is not defined the table name will be the same as \code{out_castable_sentiment} with "_features" added}

\item{key_column}{Key column name for unique identifier}

\item{document_column}{text variable column name}

\item{hostname}{\code{NULL} by default, extracts the server name from scoring code file.}

\item{astore}{If set to \code{TRUE}, it will create a scoring code that uploads the astore, available for SAS Viya 2021.1.4 or higher}

\item{astore_caslib}{Only used when \code{astore = TRUE}. The caslib where the astore model is uploaded}

\item{astore_name}{Only used when \code{astore = TRUE}. The castable name where the astore model is uploaded}

\item{astore_path}{Only used when \code{astore = TRUE}. The filepath to the astore file (extract from .zip first)}

\item{copyVars}{Only used when \code{astore = TRUE} default \code{NULL}, will not copy variables of the scoring table to the output. If \code{"ALL"} will copy all variables to the scored table output, if it is a vector, will copy named vars e.g: \verb{c("var1", "var2)}}
}
\value{
List object with the Rscore code, out castable, out caslib and the written file path.
}
\description{
It will read the score code that is written as SAS Code extract the language and hostame,
then write an R code equivalent using the \code{SWAT} package.
}
\examples{
nlp_sentiment_translate("filepath.zip")

}
