% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_translators.R
\name{nlp_topics_translate}
\alias{nlp_topics_translate}
\title{NLP Topics Translator}
\usage{
nlp_topics_translate(
  in_file = NULL,
  out_file = "topicsScoreCode.R",
  zip = TRUE,
  in_caslib,
  in_castable,
  out_caslib,
  out_castable,
  hostname = NULL,
  copyVars = NULL
)
}
\arguments{
\item{in_file}{Path to zip/sas file with the score code.}

\item{out_file}{Path to the translated \code{.R} code file.}

\item{zip}{Boolean, if the score code is inside a zip file. Default to \code{TRUE}. It will look for \code{dmcas_epscorecode.sas} file inside the zip.}

\item{in_caslib}{caslib name of the input table}

\item{in_castable}{input table name}

\item{out_caslib}{caslib name of the output table}

\item{out_castable}{name of the output table}

\item{copyVars}{default \code{NULL}, will not copy variables to the output. If \code{"ALL"} will copy all variables to the scored table output, if it is a vector, will copy named vars e.g: \verb{c("var1", "var2)}}
}
\value{
List object with the Rscore code, out castable, out caslib and the written file path.
}
\description{
It will read the score code that is written as SAS Code extract the astore and hostame information,
then write an R code equivalent in \code{SWAT}.
}
\examples{
nlp_topics_translate("filepath.zip")

}
